backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pymongo==4.10.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.0
certifi==2025.1.31
email-validator==2.2.0
transformers==4.47.1
spacy==3.8.4
torch==2.6.0
pytesseract==0.3.13
Pillow==11.0.0
PyMuPDF==1.25.1
